Prevent helm from reordering buffers after narrowing begins
At init helm, buffers are ordered by recency, However after we begin to
narrow down to select a candidate, the buffers change to being sorted
by length. This is nonsensical because more often then not we are
looking for the most recent buffer matching the pattern, not the
shortest. See https://github.com/emacs-helm/helm/issues/763 for details.