Remove parent hardlinks from helm file candidates
This is a really easy fix, but the helm maintainer is adamantly
opposed to it. See https://github.com/emacs-helm/helm/issues/943 for
more details.

The only real use we have for the hardlink is when we are reading a
file and we want to select the current directory. So we leave the
current directory hardlink "." But remove the parent hardlink "..".
However this still has the issue where each new helm directory list
has the hardlink preselect, which we don't want most of the time.
Therefore we move the cursor automatically to the second candidate.
This works much better with adaptive sorting.