Add a significant performace boost to project buffers
After doing some profiling I determined that reason
helm-browse-project-buffers was so unbearably slow is because of
calling file-truename, which took 97% of the run time. So we are going
to take a different approach where we only determine the truename
once, then save it as buffer-local variable. This took the execution
time from about 4 seconds to about 0.01 seconds. This makes this
function fast enough to actually be useable.